+++ src/compojure/api/meta.clj
@@ -1,5 +1,6 @@
 (ns compojure.api.meta
   (:require [clojure.edn :as edn]
+            [clojure.walk :as walk]
             [compojure.api.common :refer [extract-parameters]]
             [compojure.api.middleware :as mw]
             [compojure.api.routes :as routes]
@@ -295,10 +296,12 @@
       "  (ok {:name \"Kirsi\"))")))
 
 (defmethod restructure-param :return [_ schema acc]
-  (let [response (convert-return schema)]
+  (let [response (convert-return schema)
+        g (gensym 'response)]
     (-> acc
-        (update-in [:info :public :responses] (fnil conj []) response)
-        (update-in [:responses] (fnil conj []) response))))
+        (update-in [:outer-lets] into [g response])
+        (update-in [:info :public :responses] (fnil conj []) g)
+        (update-in [:responses] (fnil conj []) g))))
 
 ;;
 ;; responses
@@ -321,9 +324,11 @@
       "  (bad-request \"kosh\"))")))
 
 (defmethod restructure-param :responses [_ responses acc]
-  (-> acc
-      (update-in [:info :public :responses] (fnil conj []) responses)
-      (update-in [:responses] (fnil conj []) responses)))
+  (let [g (gensym 'responses)]
+    (-> acc
+        (update :outer-lets into [g responses])
+        (update-in [:info :public :responses] (fnil conj []) g)
+        (update-in [:responses] (fnil conj []) g))))
 
 ;;
 ;; body
@@ -338,10 +343,16 @@
       "  :body [body User]"
       "  (ok body))")))
 
-(defmethod restructure-param :body [_ [value schema] acc]
-  (-> acc
-      (update-in [:lets] into [value (src-coerce! schema :body-params :body false)])
-      (assoc-in [:info :public :parameters :body] schema)))
+(defmethod restructure-param :body [_ [value schema :as bv] acc]
+  (when-not (= "true" (System/getProperty "compojure.api.meta.allow-bad-body"))
+    (assert (= 2 (count bv))
+            (str ":body should be [sym schema], provided: " bv
+                 "\nDisable this check with -Dcompojure.api.meta.allow-bad-body=true")))
+  (let [g (gensym 'body-schema)]
+    (-> acc
+        (update :outer-lets into [g schema])
+        (update-in [:lets] into [value (src-coerce! g :body-params :body false)])
+        (assoc-in [:info :public :parameters :body] g))))
 
 ;;
 ;; query
@@ -356,10 +367,16 @@
       "  :query [params {:q s/Str, :max s/Int}]"
       "  (ok params))")))
 
-(defmethod restructure-param :query [_ [value schema] acc]
-  (-> acc
-      (update-in [:lets] into [value (src-coerce! schema :query-params :string)])
-      (assoc-in [:info :public :parameters :query] schema)))
+(defmethod restructure-param :query [_ [value schema :as bv] acc]
+  (when-not (= "true" (System/getProperty "compojure.api.meta.allow-bad-query"))
+    (assert (= 2 (count bv))
+            (str ":query should be [sym schema], provided: " bv
+                 "\nDisable this check with -Dcompojure.api.meta.allow-bad-query=true")))
+  (let [g (gensym 'query-schema)]
+    (-> acc
+        (update :outer-lets into [g schema])
+        (update-in [:lets] into [value (src-coerce! g :query-params :string)])
+        (assoc-in [:info :public :parameters :query] g))))
 
 ;;
 ;; headers
@@ -374,10 +391,16 @@
       "  :headers [headers HeaderSchema]"
       "  (ok headers))")))
 
-(defmethod restructure-param :headers [_ [value schema] acc]
-  (-> acc
-      (update-in [:lets] into [value (src-coerce! schema :headers :string)])
-      (assoc-in [:info :public :parameters :header] schema)))
+(defmethod restructure-param :headers [_ [value schema :as bv] acc]
+  (when-not (= "true" (System/getProperty "compojure.api.meta.allow-bad-headers"))
+    (assert (= 2 (count bv))
+            (str ":headers should be [sym schema], provided: " bv
+                 "\nDisable this check with -Dcompojure.api.meta.allow-bad-headers=true")))
+  (let [g (gensym 'headers-schema)]
+    (-> acc
+        (update :outer-lets into [g schema])
+        (update-in [:lets] into [value (src-coerce! g :headers :string)])
+        (assoc-in [:info :public :parameters :header] g))))
 
 ;;
 ;; body-params
@@ -393,10 +416,12 @@
       "  (ok {:total (+ x y)}))")))
 
 (defmethod restructure-param :body-params [_ body-params acc]
-  (let [schema (strict (fnk-schema body-params))]
+  (let [schema (strict (fnk-schema body-params))
+        g (gensym 'body-params-schema)]
     (-> acc
-        (update-in [:letks] into [body-params (src-coerce! schema :body-params :body)])
-        (assoc-in [:info :public :parameters :body] schema))))
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [body-params (src-coerce! g :body-params :body)])
+        (assoc-in [:info :public :parameters :body] g))))
 
 ;;
 ;; form-params
@@ -413,10 +438,12 @@
       "  (ok {:total (+ x y)}))")))
 
 (defmethod restructure-param :form-params [_ form-params acc]
-  (let [schema (strict (fnk-schema form-params))]
+  (let [schema (strict (fnk-schema form-params))
+        g (gensym 'form-params-schema)]
     (-> acc
-        (update-in [:letks] into [form-params (src-coerce! schema :form-params :string)])
-        (update-in [:info :public :parameters :formData] st/merge schema)
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [form-params (src-coerce! g :form-params :string)])
+        (update-in [:info :public :parameters :formData] #(if % (list `st/merge % g) g))
         (assoc-in [:info :public :consumes] ["application/x-www-form-urlencoded"]))))
 
 ;;
@@ -438,10 +465,12 @@
       "  (ok (dissoc foo :tempfile)))")))
 
 (defmethod restructure-param :multipart-params [_ params acc]
-  (let [schema (strict (fnk-schema params))]
+  (let [schema (strict (fnk-schema params))
+        g (gensym 'multipart-params-schema)]
     (-> acc
-        (update-in [:letks] into [params (src-coerce! schema :multipart-params :string)])
-        (update-in [:info :public :parameters :formData] st/merge schema)
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [params (src-coerce! g :multipart-params :string)])
+        (update-in [:info :public :parameters :formData] #(if % (list `st/merge % g) g))
         (assoc-in [:info :public :consumes] ["multipart/form-data"]))))
 
 ;;
@@ -458,10 +487,12 @@
       "  (ok {:total (+ x y)}))")))
 
 (defmethod restructure-param :header-params [_ header-params acc]
-  (let [schema (fnk-schema header-params)]
+  (let [schema (fnk-schema header-params)
+        g (gensym 'multipart-params-schema)]
     (-> acc
-        (update-in [:letks] into [header-params (src-coerce! schema :headers :string)])
-        (assoc-in [:info :public :parameters :header] schema))))
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [header-params (src-coerce! g :headers :string)])
+        (assoc-in [:info :public :parameters :header] g))))
 
 ;;
 ;; :query-params
@@ -477,10 +508,12 @@
       "  (ok {:total (+ x y)}))")))
 
 (defmethod restructure-param :query-params [_ query-params acc]
-  (let [schema (fnk-schema query-params)]
+  (let [schema (fnk-schema query-params)
+        g (gensym 'multipart-params-schema)]
     (-> acc
-        (update-in [:letks] into [query-params (src-coerce! schema :query-params :string)])
-        (assoc-in [:info :public :parameters :query] schema))))
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [query-params (src-coerce! g :query-params :string)])
+        (assoc-in [:info :public :parameters :query] g))))
 
 ;;
 ;; path-params
@@ -496,10 +529,12 @@
       "  (ok {:total (+ x y)}))")))
 
 (defmethod restructure-param :path-params [_ path-params acc]
-  (let [schema (fnk-schema path-params)]
+  (let [schema (fnk-schema path-params)
+        g (gensym 'form-params-schema)]
     (-> acc
-        (update-in [:letks] into [path-params (src-coerce! schema :route-params :string)])
-        (assoc-in [:info :public :parameters :path] schema))))
+        (update :outer-lets into [g schema])
+        (update-in [:letks] into [path-params (src-coerce! g :route-params :string)])
+        (assoc-in [:info :public :parameters :path] g))))
 
 ;;
 ;; middleware
@@ -568,7 +603,7 @@
 
 (defmethod help/help-for [:meta :coercion] [_ _]
   (help/text
-    "Route-spesific overrides for coercion. See more on wiki:"
+    "Route-specific overrides for coercion. See more on wiki:"
     "https://github.com/metosin/compojure-api/wiki/Validation-and-coercion\n"
     (help/code
       "(POST \"/user\" []"
@@ -577,22 +612,24 @@
       "  (ok user))")))
 
 (defmethod restructure-param :coercion [_ coercion acc]
-  (-> acc
-      (assoc-in [:info :coercion] coercion)
-      (update-in [:middleware] conj [`mw/wrap-coercion coercion])))
+  (let [g (gensym 'coercion)]
+    (-> acc
+        (update :outer-lets into [g coercion])
+        (assoc-in [:info :coercion] g)
+        (update-in [:middleware] conj [`mw/wrap-coercion g]))))
 
 ;;
 ;; Impl
 ;;
 
 (defmacro dummy-let
-  "Dummy let-macro used in resolving route-docs. not part of normal invokation chain."
+  "Dummy let-macro used in resolving route-docs. not part of normal invocation chain."
   [bindings & body]
   (let [bind-form (vec (apply concat (for [n (take-nth 2 bindings)] [n nil])))]
     `(let ~bind-form ~@body)))
 
 (defmacro dummy-letk
-  "Dummy letk-macro used in resolving route-docs. not part of normal invokation chain."
+  "Dummy letk-macro used in resolving route-docs. not part of normal invocation chain."
   [bindings & body]
   (reduce
     (fn [cur-body-form [bind-form]]
@@ -612,7 +649,8 @@
   [path route]
   `(compojure.api.compojure-compat/make-context
      ~(#'compojure.core/context-route path)
-     (constantly ~route)))
+     (let [r# ~route]
+       (fn [_#] r#))))
 
 (defn routing [handlers]
   (if-let [handlers (seq (keep identity (flatten handlers)))]
@@ -731,24 +769,7 @@
                (when (var? v)
                  (when (middleware-vars (symbol v))
                    (let [[_ path route-arg & args] body
-                         [options body] (extract-parameters args true)
-                         [path-string lets arg-with-request] (destructure-compojure-api-request path route-arg)
-                         {:keys [lets
-                                 letks
-                                 responses
-                                 middleware
-                                 info
-                                 swagger
-                                 body]} (reduce
-                                          (fn [acc [k v]]
-                                            (restructure-param k v (update-in acc [:parameters] dissoc k)))
-                                          {:lets lets
-                                           :letks []
-                                           :responses nil
-                                           :middleware []
-                                           :info {}
-                                           :body body}
-                                          options)]
+                         [options body] (extract-parameters args true)]
                      (static-body? &env body))))))))))
 
 (def route-middleware-vars (into #{}
@@ -921,6 +942,7 @@
 
         {:keys [lets
                 letks
+                outer-lets
                 responses
                 middleware
                 info
@@ -930,6 +952,7 @@
                            (restructure-param k v (update-in acc [:parameters] dissoc k)))
                          {:lets lets
                           :letks []
+                          :outer-lets [] ;; lets around the call to map->Route
                           :responses nil
                           :middleware []
                           :info {}
@@ -942,6 +965,7 @@
                             (-> info :public :static)))
                   "Cannot be both a :dynamic and :static context.")
 
+        ;; I think it's ok if we have :outer-lets
         bindings? (boolean (or (route-args? route-arg) (seq lets) (seq letks)))
 
         _ (assert (not (and (-> info :public :static)
@@ -1064,24 +1088,26 @@
                          form `(compojure.core/let-request [~arg-with-request ~'+compojure-api-request+] ~form)
                          form `(fn [~'+compojure-api-request+] ~form)
                          form `(delay (flatten (~form {})))]
-                     form)]
-
-        `(routes/map->Route
-           {:path ~path-string
-            :method ~method
-            :info (merge-parameters ~info)
-            :childs ~childs
-            :handler ~form}))
+                     form)
+            form `(routes/map->Route
+                    {:path ~path-string
+                     :method ~method
+                     :info (merge-parameters ~info)
+                     :childs ~childs
+                     :handler ~form})
+            form (if (seq outer-lets) `(let ~outer-lets ~form) form)]
+        form)
 
       ;; endpoints
       (let [form `(do ~@body)
             form (if (seq letks) `(p/letk ~letks ~form) form)
             form (if (seq lets) `(let ~lets ~form) form)
             form (compojure.core/compile-route method path arg-with-request (list form))
-            form (if (seq middleware) `(compojure.core/wrap-routes ~form (mw/compose-middleware ~middleware)) form)]
-
-        `(routes/map->Route
-           {:path ~path-string
-            :method ~method
-            :info (merge-parameters ~info)
-            :handler ~form})))))
+            form (if (seq middleware) `(compojure.core/wrap-routes ~form (mw/compose-middleware ~middleware)) form)
+            form `(routes/map->Route
+                    {:path ~path-string
+                     :method ~method
+                     :info (merge-parameters ~info)
+                     :handler ~form})
+            form (if (seq outer-lets) `(let ~outer-lets ~form) form)]
+        form))))